@@ -1,25 +1,30 @@
 <template>
-  <div id="app">
-    <ProductListOne></ProductListOne>
-    <ProductListTwo></ProductListTwo>
-  </div>
+  <nav>
+    <router-link to="/">Home</router-link> |
+    <router-link to="/about">About</router-link>
+  </nav>
+  <router-view/>
 </template>
 
-<script>
-import ProductListOne from "./components/ProductListOne.vue"
-import ProductListTwo from "./components/ProductListTwo.vue"
-export default {
-    name: "app",
-    components: {
-        ProductListOne,
-        ProductListTwo
-    },
-    data() {
-        return {};
-    },
-}
-</script>
-
 <style>
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+}
 
+nav {
+  padding: 30px;
+}
+
+nav a {
+  font-weight: bold;
+  color: #2c3e50;
+}
+
+nav a.router-link-exact-active {
+  color: #42b983;
+}
 </style>
